% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-socialbladeR.Rmd.
\name{top_50}
\alias{top_50}
\title{Get top 50 followed accounts from either youtube, twitch.tv, facebook}
\usage{
top_50(platform)
}
\arguments{
\item{platform}{Name of platform beings searched}
}
\description{
Get top 50 followed accounts from either youtube, twitch.tv, facebook
}