% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-socialbladeR.Rmd.
\name{yt_topcategories}
\alias{yt_topcategories}
\title{Say hello to someone}
\usage{
yt_topcategories(keyword)
}
\arguments{
\item{keyword}{Keyword is the type of category being searched}
}
\description{
Say hello to someone
}